feat(utils): add appendStylesheetTo dom helper
Moved from core's dom-helpers so client can reuse it instead of
hand-rolling the same <link rel=stylesheet> construction.
